Updated RobotSphere (player)
Modified the Robot Animation Script to change animation states. Removed the free robot anim script and just copied the script onto the new one I created (I didn't know how to adjust the editor). I didn't want to mess with the editor script since it is late as of writing this and I don't want to stay up any later.